Update music_cog.py
I have made improvements with the search function, idc about credits so you can use this code as an example.  I have left a decent amount of commentary so please check it out.  Lets work together as a community to make an open source music bot!!!  using the cog setup is genius as many could implement it without editing it once this code is top notch.  I would recommend added a requirments.txt with the python modules and versions of them so this can be ran as long the versions don't get erased from the web.